# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "HaiTouWang Cloud API",
    "version": "1.0.0",
    "description": "Contract-first API for profile, job, material, review, local agent, application and lifecycle domains. Raw target-site credentials, cookies and CAPTCHA values are forbidden in every cloud request.",
    "license": {
      "name": "Proprietary",
      "identifier": "LicenseRef-Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.haitouwang.example",
      "description": "Production placeholder; region remains a product decision"
    }
  ],
  "security": [
    {
      "UserBearer": []
    }
  ],
...
  "paths": {
...
    "/v1/materials/{material_id}/approve": {
      "post": {
        "tags": ["Materials"],
        "operationId": "approveMaterial",
        "summary": "Approve a material after a server-side Review gate",
        "description": "The authenticated principal supplies no tenant identity. The server reads the persisted Review and material in one consistency boundary and rejects missing, stale, non-approved, or open must-fix Reviews.",
        "x-required-user-permission": "material:approve",
        "x-audit-event": "material.approved_or_rejected",
        "parameters": [
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ApproveMaterialRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Material approved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Material" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "412": { "$ref": "#/components/responses/PreconditionFailed" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/materials/{material_id}/reject": {
      "post": {
        "tags": ["Materials"],
        "operationId": "rejectMaterial",
        "summary": "Reject a material with an audited server-side transition",
        "x-required-user-permission": "material:approve",
        "x-audit-event": "material.rejected_or_rejection_rejected",
        "parameters": [
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RejectMaterialRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Material rejected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Material" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/v1/reviews/{review_id}": {
      "get": {
        "tags": ["Reviews"],
        "operationId": "getReview",
        "summary": "Get a tenant-owned persisted Review",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Persisted Review",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Review" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/reviews/{review_id}/findings/{finding_id}/resolve": {
      "post": {
        "tags": ["Reviews"],
        "operationId": "resolveReviewFinding",
        "summary": "Resolve a Review finding before material approval",
        "x-required-user-permission": "material:approve",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Uuid" }
          },
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated Review",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Review" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/v1/audit-events": {
      "get": {
        "tags": ["Audit"],
        "operationId": "listAuditEvents",
        "summary": "List tenant-visible append-only audit events",
        "parameters": [
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/PageSize" },
          {
            "name": "resource_id",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/Uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor page of redacted audit events",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuditEventPage" }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "UserBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived SaaS user access token; tenant is derived from the authenticated principal."
      },
      "AgentBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Five-minute device-bound agent token with least-privilege scope claims. Every request also carries proof of possession and a replay-protected nonce."
      }
    },
    "parameters": {
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "description": "Opaque cursor bound to tenant, filters, ordering and API major version.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 2048
        }
      },
      "PageSize": {
        "name": "page_size",
        "in": "query",
        "description": "Number of items to return.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 25
        }
      },
...
    },
    "headers": {
      "RequestId": {
        "description": "Server-generated identifier for this request.",
        "schema": { "$ref": "#/components/schemas/Uuid" }
      },
...
    },
    "responses": {
      "Error": {
        "description": "Error response using a stable machine code",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/RequestId" }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Authentication required, expired or revoked",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "Forbidden": {
        "description": "Tenant, policy or token scope forbids the action",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found in the authenticated tenant",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "Conflict": {
        "description": "Version, unique key, idempotency key or state transition conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "PreconditionFailed": {
        "description": "Explicit confirmation or verified evidence is required",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
...
    },
    "schemas": {
      "Uuid": {
        "type": "string",
        "format": "uuid"
      },
...
      "Fact": { "$ref": "../schemas/domain.schema.json#/$defs/Fact" },
...
      "Material": { "$ref": "../schemas/domain.schema.json#/$defs/Material" },
...
      "Review": { "$ref": "../schemas/domain.schema.json#/$defs/Review" },
...
      "AuditEvent": { "$ref": "../schemas/domain.schema.json#/$defs/AuditEvent" },
      "PageMeta": { "$ref": "../schemas/domain.schema.json#/$defs/PageMeta" },
      "ErrorEnvelope": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorEnvelope" },
...
      "AuditEventPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "page"],
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AuditEvent" }
          },
          "page": { "$ref": "#/components/schemas/PageMeta" }
        }
      },
...
      "ApproveMaterialRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expected_version", "facts", "evaluated_at", "review_id"],
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1
          },
          "facts": {
            "type": "array",
            "maxItems": 500,
            "items": { "$ref": "#/components/schemas/Fact" }
          },
          "evaluated_at": {
            "type": "string",
            "format": "date-time"
          },
          "goal_id": { "$ref": "#/components/schemas/Uuid" },
          "review_id": { "$ref": "#/components/schemas/Uuid" }
        }
      },
      "RejectMaterialRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expected_version"],
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
...
    }
  }
}